// src/app/services/models/reclamation.ts
/* tslint:disable */
/* eslint-disable */
export interface Reclamation {
  contenu?: string;
  createdBy?: string;
  createdDate?: string;
  etat?: 'En cours de traitement' | 'Trait\xE9';
  firstname?: string;
  gmail?: string;
  idReclamation?: number;
  lastname?: string;
  reponse?: string;
  type?: 'ProblemeEnchere' | 'ProblemePaiement' | 'Problemecompteutilisateur' | 'Problemetechnique';
}